deploy/common/requirements.txt
mem0ai==1.0.5
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-dotenv==1.2.2
sentence-transformers==5.1.2
